Move EditorPaneVis to new component
